extend unit testing on lib sensirion

+++ lib/sensirion/sensirion.hpp
@@ -13,11 +13,10 @@ class sensirion {
     static bool checkCrc(const uint8_t* data, uint32_t count);           // check all CRCs in a buffer of data. Every third byte is a CRC of the previous two bytes.
     static void insertCrc(uint8_t* data, uint32_t count);                // insert CRCs in a buffer of data. Every third byte is a CRC of the previous two bytes.
     static float asFloat(const uint8_t* bytes);                          // convert a sensirion 4-byte array to a float
-    static float asUint32(const uint8_t* bytes);                         // convert a sensirion 4-byte array to a uint32_t
-    static float asUint16(const uint8_t* bytes);                         // convert a sensirion 2-byte array to a uint16_t
+    static uint32_t asUint32(const uint8_t* bytes);                      // convert a sensirion 4-byte array to a uint32_t
+    static uint16_t asUint16(const uint8_t* bytes);                      // convert a sensirion 2-byte array to a uint16_t
     static constexpr uint8_t crcPolynome{0x31};
     static constexpr uint8_t crcInit{0xFF};
-
 
 #ifndef unitTesting
